% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbbfunction.R
\name{acircle}
\alias{acircle}
\title{Area of circle}
\usage{
acircle(r)
}
\arguments{
\item{r}{is the radius of the cirlce}
}
\value{
The area of circle
}
\description{
Compute the area of circle shape
}
\examples{
area1 <- acircle(4);
area2 <- acircle(3);
}
