% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbbfunction.R
\name{atrapezoid}
\alias{atrapezoid}
\title{Area of trapezoid}
\usage{
atrapezoid(b1, b2, h)
}
\arguments{
\item{b1}{is the length of first parallel side}

\item{b2}{is the length of second parallel side}

\item{h}{is the distance (height) between the parallels}
}
\value{
The area of trapezoid
}
\description{
Compute the area of trapezoid shape
}
\examples{
area1 <- atrapezoid(5,2,4);
area2 <- atrapezoid(8,3,1);
}
