% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbbfunction.R
\name{psquare}
\alias{psquare}
\title{Perimeter of square}
\usage{
psquare(s)
}
\arguments{
\item{s}{is the length of square}
}
\value{
The perimeter of square
}
\description{
Compute the perimeter of square shape
}
\examples{
perimeter1 <- psquare(5);
perimeter2 <- psquare(8);
}
