% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbbfunction.R
\name{ptriangle}
\alias{ptriangle}
\title{Perimeter of triangle}
\usage{
ptriangle(a, b, c)
}
\arguments{
\item{a}{is the first side length}

\item{b}{is the second side length}

\item{c}{is the third side length}
}
\value{
The perimeter of triangle
}
\description{
Compute the perimeter of triangle shape
}
\examples{
perimeter1 <- ptriangle(5,2,4);
perimeter2 <- ptriangle(8,3,1);
}
