% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbbfunction.R
\name{vpyramid}
\alias{vpyramid}
\title{Volume of pyramid}
\usage{
vpyramid(a, h)
}
\arguments{
\item{a}{is the area of the base}

\item{h}{is the height of pyramid}
}
\value{
The volume of pyramid
}
\description{
Compute the volume of pyramid shape
}
\examples{
volume1 <- vpyramid(5,2);
volume2 <- vpyramid(3,1);
}
