% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbbfunction.R
\name{vrrectangle}
\alias{vrrectangle}
\title{Volume of rectangular}
\usage{
vrrectangle(l, w, h)
}
\arguments{
\item{l}{is the length rectangle}

\item{w}{is the width of rectangle}

\item{h}{is the height of rectangle}
}
\value{
The volume of rectangular
}
\description{
Compute the volume of rectangular shape
}
\examples{
volume1 <- vrrectangle(5,2,4);
volume2 <- vrrectangle(8,3,1);
}
